@@ -36,11 +36,10 @@
 
 #include <string.h>
 
-#include <boost/array.hpp>
-#include <boost/chrono.hpp>
-#include <boost/cstdint.hpp>
-#include <boost/limits.hpp>
+#include <array>
+#include <cstdint>
 #include <map>
+#include <memory>
 #include <new>
 #include <queue>
 #include <sstream>
@@ -142,8 +141,8 @@ Channel::OpenOpts Channel::OpenOpts::FromUri(const std::string &uri) {
   amqp_connection_info info;
   amqp_default_connection_info(&info);
 
-  boost::shared_ptr<char> uri_dup =
-      boost::shared_ptr<char>(strdup(uri.c_str()), free);
+  std::shared_ptr<char> uri_dup =
+      std::shared_ptr<char>(strdup(uri.c_str()), free);
 
   if (0 != amqp_parse_url(uri_dup.get(), &info)) {
     throw BadUriException();
@@ -193,22 +192,22 @@ Channel::ptr_t Channel::Open(const OpenOpts &opts) {
     throw std::runtime_error(
         "opts.port is not valid, it must be a positive number");
   }
-  if (opts.auth.empty()) {
+  if (opts.auth.index()==0) {
     throw std::runtime_error("opts.auth is not specified, it is required");
   }
-  if (!opts.tls_params.is_initialized()) {
-    switch (opts.auth.which()) {
-      case 0: {
+  if (!opts.tls_params.has_value()) {
+    switch (opts.auth.index()) {
+      case 1: {
         const OpenOpts::BasicAuth &auth =
-            boost::get<OpenOpts::BasicAuth>(opts.auth);
-        return boost::make_shared<Channel>(
+            std::get<OpenOpts::BasicAuth>(opts.auth);
+        return std::make_shared<Channel>(
             OpenChannel(opts.host, opts.port, auth.username, auth.password,
                         opts.vhost, opts.frame_max, false));
       }
-      case 1: {
+      case 2: {
         const OpenOpts::ExternalSaslAuth &auth =
-            boost::get<OpenOpts::ExternalSaslAuth>(opts.auth);
-        return boost::make_shared<Channel>(
+            std::get<OpenOpts::ExternalSaslAuth>(opts.auth);
+        return std::make_shared<Channel>(
             OpenChannel(opts.host, opts.port, auth.identity, "", opts.vhost,
                         opts.frame_max, true));
       }
@@ -216,20 +215,20 @@ Channel::ptr_t Channel::Open(const OpenOpts &opts) {
         throw std::logic_error("Unhandled auth type");
     }
   }
-  switch (opts.auth.which()) {
-    case 0: {
-      const OpenOpts::BasicAuth &auth =
-          boost::get<OpenOpts::BasicAuth>(opts.auth);
-      return boost::make_shared<Channel>(OpenSecureChannel(
-          opts.host, opts.port, auth.username, auth.password, opts.vhost,
-          opts.frame_max, opts.tls_params.get(), false));
-    }
+  switch (opts.auth.index()) {
     case 1: {
+      const OpenOpts::BasicAuth &auth =
+          std::get<OpenOpts::BasicAuth>(opts.auth);
+      return std::make_shared<Channel>(OpenSecureChannel(
+          opts.host, opts.port, auth.username, auth.password, opts.vhost,
+          opts.frame_max, opts.tls_params.value(), false));
+    }
+    case 2: {
       const OpenOpts::ExternalSaslAuth &auth =
-          boost::get<OpenOpts::ExternalSaslAuth>(opts.auth);
-      return boost::make_shared<Channel>(
+          std::get<OpenOpts::ExternalSaslAuth>(opts.auth);
+      return std::make_shared<Channel>(
           OpenSecureChannel(opts.host, opts.port, auth.identity, "", opts.vhost,
-                            opts.frame_max, opts.tls_params.get(), true));
+                            opts.frame_max, opts.tls_params.value(), true));
     }
     default:
       throw std::logic_error("Unhandled auth type");
@@ -340,7 +339,7 @@ Channel::ptr_t Channel::CreateSecureSaslExternal(
 
 Channel::ptr_t Channel::CreateFromUri(const std::string &uri, int frame_max) {
   OpenOpts opts = OpenOpts::FromUri(uri);
-  if (opts.tls_params.is_initialized()) {
+  if (opts.tls_params.has_value()) {
     throw std::runtime_error(
         "CreateFromUri only supports non-SSL-enabled URIs");
   }
@@ -354,7 +353,7 @@ Channel::ptr_t Channel::CreateSecureFromUri(
     const std::string &path_to_client_cert, bool verify_hostname_and_peer,
     int frame_max) {
   OpenOpts opts = OpenOpts::FromUri(uri);
-  if (!opts.tls_params.is_initialized()) {
+  if (!opts.tls_params.has_value()) {
     throw std::runtime_error(
         "CreateSecureFromUri only supports SSL-enabled URIs");
   }
@@ -477,9 +476,9 @@ int Channel::GetSocketFD() const {
   return amqp_get_sockfd(m_impl->m_connection);
 }
 
-bool Channel::CheckExchangeExists(boost::string_ref exchange_name) {
-  const boost::array<boost::uint32_t, 1> DECLARE_OK = {
-      {AMQP_EXCHANGE_DECLARE_OK_METHOD}};
+bool Channel::CheckExchangeExists(std::string_view exchange_name) {
+  const std::array<std::uint32_t, 1> DECLARE_OK = {
+      AMQP_EXCHANGE_DECLARE_OK_METHOD};
 
   amqp_exchange_declare_t declare = {};
   declare.exchange = StringRefToBytes(exchange_name);
@@ -507,8 +506,8 @@ void Channel::DeclareExchange(const std::string &exchange_name,
                               const std::string &exchange_type, bool passive,
                               bool durable, bool auto_delete,
                               const Table &arguments) {
-  const boost::array<boost::uint32_t, 1> DECLARE_OK = {
-      {AMQP_EXCHANGE_DECLARE_OK_METHOD}};
+  const std::array<std::uint32_t, 1> DECLARE_OK = {
+      AMQP_EXCHANGE_DECLARE_OK_METHOD};
   m_impl->CheckIsConnected();
 
   amqp_exchange_declare_t declare = {};
@@ -530,8 +529,8 @@ void Channel::DeclareExchange(const std::string &exchange_name,
 }
 
 void Channel::DeleteExchange(const std::string &exchange_name, bool if_unused) {
-  const boost::array<boost::uint32_t, 1> DELETE_OK = {
-      {AMQP_EXCHANGE_DELETE_OK_METHOD}};
+  const std::array<std::uint32_t, 1> DELETE_OK = {
+      AMQP_EXCHANGE_DELETE_OK_METHOD};
   m_impl->CheckIsConnected();
 
   amqp_exchange_delete_t del = {};
@@ -554,8 +553,7 @@ void Channel::BindExchange(const std::string &destination,
                            const std::string &source,
                            const std::string &routing_key,
                            const Table &arguments) {
-  const boost::array<boost::uint32_t, 1> BIND_OK = {
-      {AMQP_EXCHANGE_BIND_OK_METHOD}};
+  const std::array<std::uint32_t, 1> BIND_OK = {AMQP_EXCHANGE_BIND_OK_METHOD};
   m_impl->CheckIsConnected();
 
   amqp_exchange_bind_t bind = {};
@@ -582,8 +580,8 @@ void Channel::UnbindExchange(const std::string &destination,
                              const std::string &source,
                              const std::string &routing_key,
                              const Table &arguments) {
-  const boost::array<boost::uint32_t, 1> UNBIND_OK = {
-      {AMQP_EXCHANGE_UNBIND_OK_METHOD}};
+  const std::array<std::uint32_t, 1> UNBIND_OK = {
+      AMQP_EXCHANGE_UNBIND_OK_METHOD};
   m_impl->CheckIsConnected();
 
   amqp_exchange_unbind_t unbind = {};
@@ -601,9 +599,9 @@ void Channel::UnbindExchange(const std::string &destination,
   m_impl->MaybeReleaseBuffersOnChannel(frame.channel);
 }
 
-bool Channel::CheckQueueExists(boost::string_ref queue_name) {
-  const boost::array<boost::uint32_t, 1> DECLARE_OK = {
-      {AMQP_QUEUE_DECLARE_OK_METHOD}};
+bool Channel::CheckQueueExists(std::string_view queue_name) {
+  const std::array<std::uint32_t, 1> DECLARE_OK = {
+      AMQP_QUEUE_DECLARE_OK_METHOD};
 
   amqp_queue_declare_t declare = {};
   declare.queue = StringRefToBytes(queue_name);
@@ -630,16 +628,16 @@ std::string Channel::DeclareQueue(const std::string &queue_name, bool passive,
 std::string Channel::DeclareQueue(const std::string &queue_name, bool passive,
                                   bool durable, bool exclusive,
                                   bool auto_delete, const Table &arguments) {
-  boost::uint32_t message_count;
-  boost::uint32_t consumer_count;
+  std::uint32_t message_count;
+  std::uint32_t consumer_count;
   return DeclareQueueWithCounts(queue_name, message_count, consumer_count,
                                 passive, durable, exclusive, auto_delete,
                                 arguments);
 }
 
 std::string Channel::DeclareQueueWithCounts(const std::string &queue_name,
-                                            boost::uint32_t &message_count,
-                                            boost::uint32_t &consumer_count,
+                                            std::uint32_t &message_count,
+                                            std::uint32_t &consumer_count,
                                             bool passive, bool durable,
                                             bool exclusive, bool auto_delete) {
   return DeclareQueueWithCounts(queue_name, message_count, consumer_count,
@@ -648,13 +646,13 @@ std::string Channel::DeclareQueueWithCounts(const std::string &queue_name,
 }
 
 std::string Channel::DeclareQueueWithCounts(const std::string &queue_name,
-                                            boost::uint32_t &message_count,
-                                            boost::uint32_t &consumer_count,
+                                            std::uint32_t &message_count,
+                                            std::uint32_t &consumer_count,
                                             bool passive, bool durable,
                                             bool exclusive, bool auto_delete,
                                             const Table &arguments) {
-  const boost::array<boost::uint32_t, 1> DECLARE_OK = {
-      {AMQP_QUEUE_DECLARE_OK_METHOD}};
+  const std::array<std::uint32_t, 1> DECLARE_OK = {
+      AMQP_QUEUE_DECLARE_OK_METHOD};
   m_impl->CheckIsConnected();
 
   amqp_queue_declare_t declare = {};
@@ -686,8 +684,7 @@ std::string Channel::DeclareQueueWithCounts(const std::string &queue_name,
 
 void Channel::DeleteQueue(const std::string &queue_name, bool if_unused,
                           bool if_empty) {
-  const boost::array<boost::uint32_t, 1> DELETE_OK = {
-      {AMQP_QUEUE_DELETE_OK_METHOD}};
+  const std::array<std::uint32_t, 1> DELETE_OK = {AMQP_QUEUE_DELETE_OK_METHOD};
   m_impl->CheckIsConnected();
 
   amqp_queue_delete_t del = {};
@@ -710,8 +707,7 @@ void Channel::BindQueue(const std::string &queue_name,
                         const std::string &exchange_name,
                         const std::string &routing_key,
                         const Table &arguments) {
-  const boost::array<boost::uint32_t, 1> BIND_OK = {
-      {AMQP_QUEUE_BIND_OK_METHOD}};
+  const std::array<std::uint32_t, 1> BIND_OK = {AMQP_QUEUE_BIND_OK_METHOD};
   m_impl->CheckIsConnected();
 
   amqp_queue_bind_t bind = {};
@@ -738,8 +734,7 @@ void Channel::UnbindQueue(const std::string &queue_name,
                           const std::string &exchange_name,
                           const std::string &routing_key,
                           const Table &arguments) {
-  const boost::array<boost::uint32_t, 1> UNBIND_OK = {
-      {AMQP_QUEUE_UNBIND_OK_METHOD}};
+  const std::array<std::uint32_t, 1> UNBIND_OK = {AMQP_QUEUE_UNBIND_OK_METHOD};
   m_impl->CheckIsConnected();
 
   amqp_queue_unbind_t unbind = {};
@@ -757,8 +752,7 @@ void Channel::UnbindQueue(const std::string &queue_name,
 }
 
 void Channel::PurgeQueue(const std::string &queue_name) {
-  const boost::array<boost::uint32_t, 1> PURGE_OK = {
-      {AMQP_QUEUE_PURGE_OK_METHOD}};
+  const std::array<std::uint32_t, 1> PURGE_OK = {AMQP_QUEUE_PURGE_OK_METHOD};
   m_impl->CheckIsConnected();
 
   amqp_queue_purge_t purge = {};
@@ -845,11 +839,10 @@ void Channel::BasicPublish(const std::string &exchange_name,
   // - channel.close - probably tried to publish to a non-existant exchange, in
   // any case error!
   // - connection.clsoe - something really bad happened
-  const boost::array<boost::uint32_t, 3> PUBLISH_ACK = {
-      {AMQP_BASIC_ACK_METHOD, AMQP_BASIC_RETURN_METHOD,
-       AMQP_BASIC_NACK_METHOD}};
+  const std::array<std::uint32_t, 3> PUBLISH_ACK = {
+      AMQP_BASIC_ACK_METHOD, AMQP_BASIC_RETURN_METHOD, AMQP_BASIC_NACK_METHOD};
   amqp_frame_t response;
-  boost::array<amqp_channel_t, 1> channels = {{channel}};
+  std::array<amqp_channel_t, 1> channels = {channel};
   m_impl->GetMethodOnChannel(channels, response, PUBLISH_ACK);
 
   if (AMQP_BASIC_NACK_METHOD == response.payload.method.id) {
@@ -868,8 +861,7 @@ void Channel::BasicPublish(const std::string &exchange_name,
                 response.payload.method.decoded)),
             channel);
 
-    const boost::array<boost::uint32_t, 1> BASIC_ACK = {
-        {AMQP_BASIC_ACK_METHOD}};
+    const std::array<std::uint32_t, 1> BASIC_ACK = {AMQP_BASIC_ACK_METHOD};
     m_impl->GetMethodOnChannel(channels, response, BASIC_ACK);
     m_impl->ReturnChannel(channel);
     m_impl->MaybeReleaseBuffersOnChannel(channel);
@@ -882,8 +874,8 @@ void Channel::BasicPublish(const std::string &exchange_name,
 
 bool Channel::BasicGet(Envelope::ptr_t &envelope, const std::string &queue,
                        bool no_ack) {
-  const boost::array<boost::uint32_t, 2> GET_RESPONSES = {
-      {AMQP_BASIC_GET_OK_METHOD, AMQP_BASIC_GET_EMPTY_METHOD}};
+  const std::array<std::uint32_t, 2> GET_RESPONSES = {
+      AMQP_BASIC_GET_OK_METHOD, AMQP_BASIC_GET_EMPTY_METHOD};
   m_impl->CheckIsConnected();
 
   amqp_basic_get_t get = {};
@@ -902,7 +894,7 @@ bool Channel::BasicGet(Envelope::ptr_t &envelope, const std::string &queue,
 
   amqp_basic_get_ok_t *get_ok =
       (amqp_basic_get_ok_t *)response.payload.method.decoded;
-  boost::uint64_t delivery_tag = get_ok->delivery_tag;
+  std::uint64_t delivery_tag = get_ok->delivery_tag;
   bool redelivered = (get_ok->redelivered == 0 ? false : true);
   std::string exchange((char *)get_ok->exchange.bytes, get_ok->exchange.len);
   std::string routing_key((char *)get_ok->routing_key.bytes,
@@ -918,8 +910,8 @@ bool Channel::BasicGet(Envelope::ptr_t &envelope, const std::string &queue,
 }
 
 void Channel::BasicRecover(const std::string &consumer) {
-  const boost::array<boost::uint32_t, 1> RECOVER_OK = {
-      {AMQP_BASIC_RECOVER_OK_METHOD}};
+  const std::array<std::uint32_t, 1> RECOVER_OK = {
+      AMQP_BASIC_RECOVER_OK_METHOD};
   m_impl->CheckIsConnected();
 
   amqp_basic_recover_t recover = {};
@@ -935,21 +927,21 @@ void Channel::BasicRecover(const std::string &consumer) {
 std::string Channel::BasicConsume(const std::string &queue,
                                   const std::string &consumer_tag,
                                   bool no_local, bool no_ack, bool exclusive,
-                                  boost::uint16_t message_prefetch_count) {
+                                  std::uint16_t message_prefetch_count) {
   return BasicConsume(queue, consumer_tag, no_local, no_ack, exclusive,
                       message_prefetch_count, Table());
 }
 std::string Channel::BasicConsume(const std::string &queue,
                                   const std::string &consumer_tag,
                                   bool no_local, bool no_ack, bool exclusive,
-                                  boost::uint16_t message_prefetch_count,
+                                  std::uint16_t message_prefetch_count,
                                   const Table &arguments) {
   m_impl->CheckIsConnected();
   amqp_channel_t channel = m_impl->GetChannel();
 
   // Set this before starting the consume as it may have been set by a previous
   // consumer
-  const boost::array<boost::uint32_t, 1> QOS_OK = {{AMQP_BASIC_QOS_OK_METHOD}};
+  const std::array<std::uint32_t, 1> QOS_OK = {AMQP_BASIC_QOS_OK_METHOD};
 
   amqp_basic_qos_t qos = {};
   qos.prefetch_size = 0;
@@ -959,8 +951,8 @@ std::string Channel::BasicConsume(const std::string &queue,
   m_impl->DoRpcOnChannel(channel, AMQP_BASIC_QOS_METHOD, &qos, QOS_OK);
   m_impl->MaybeReleaseBuffersOnChannel(channel);
 
-  const boost::array<boost::uint32_t, 1> CONSUME_OK = {
-      {AMQP_BASIC_CONSUME_OK_METHOD}};
+  const std::array<std::uint32_t, 1> CONSUME_OK = {
+      AMQP_BASIC_CONSUME_OK_METHOD};
 
   amqp_basic_consume_t consume = {};
   consume.queue = StringToBytes(queue);
@@ -989,11 +981,11 @@ std::string Channel::BasicConsume(const std::string &queue,
 }
 
 void Channel::BasicQos(const std::string &consumer_tag,
-                       boost::uint16_t message_prefetch_count) {
+                       std::uint16_t message_prefetch_count) {
   m_impl->CheckIsConnected();
   amqp_channel_t channel = m_impl->GetConsumerChannel(consumer_tag);
 
-  const boost::array<boost::uint32_t, 1> QOS_OK = {{AMQP_BASIC_QOS_OK_METHOD}};
+  const std::array<std::uint32_t, 1> QOS_OK = {AMQP_BASIC_QOS_OK_METHOD};
 
   amqp_basic_qos_t qos = {};
   qos.prefetch_size = 0;
@@ -1008,8 +1000,7 @@ void Channel::BasicCancel(const std::string &consumer_tag) {
   m_impl->CheckIsConnected();
   amqp_channel_t channel = m_impl->GetConsumerChannel(consumer_tag);
 
-  const boost::array<boost::uint32_t, 1> CANCEL_OK = {
-      {AMQP_BASIC_CANCEL_OK_METHOD}};
+  const std::array<std::uint32_t, 1> CANCEL_OK = {AMQP_BASIC_CANCEL_OK_METHOD};
 
   amqp_basic_cancel_t cancel = {};
   cancel.consumer_tag = StringToBytes(consumer_tag);
@@ -1051,7 +1042,7 @@ bool Channel::BasicConsumeMessage(const std::string &consumer_tag,
   m_impl->CheckIsConnected();
   amqp_channel_t channel = m_impl->GetConsumerChannel(consumer_tag);
 
-  boost::array<amqp_channel_t, 1> channels = {{channel}};
+  std::array<amqp_channel_t, 1> channels = {channel};
 
   return m_impl->ConsumeMessageOnChannel(channels, message, timeout);
 }
