@@ -37,9 +37,12 @@
 #include <sys/types.h>
 #endif
 
-#include <boost/algorithm/string/classification.hpp>
-#include <boost/algorithm/string/split.hpp>
-#include <boost/array.hpp>
+#include <string.h>
+
+#include <array>
+#include <cassert>
+#include <chrono>
+#include <string>
 
 #include "SimpleAmqpClient/AmqpException.h"
 #include "SimpleAmqpClient/AmqpLibraryException.h"
@@ -48,11 +51,6 @@
 #include "SimpleAmqpClient/ConnectionClosedException.h"
 #include "SimpleAmqpClient/ConsumerTagNotFoundException.h"
 #include "SimpleAmqpClient/TableImpl.h"
-#define BOOST_BIND_GLOBAL_PLACEHOLDERS
-#include <string.h>
-
-#include <boost/bind.hpp>
-#include <boost/lexical_cast.hpp>
 
 #define BROKER_HEARTBEAT 0
 
@@ -179,16 +177,16 @@ amqp_channel_t Channel::ChannelImpl::GetNextChannelId() {
 amqp_channel_t Channel::ChannelImpl::CreateNewChannel() {
   amqp_channel_t new_channel = GetNextChannelId();
 
-  static const boost::array<boost::uint32_t, 1> OPEN_OK = {
-      {AMQP_CHANNEL_OPEN_OK_METHOD}};
+  static const std::array<std::uint32_t, 1> OPEN_OK = {
+      AMQP_CHANNEL_OPEN_OK_METHOD};
   amqp_channel_open_t channel_open = {};
-  DoRpcOnChannel<boost::array<boost::uint32_t, 1> >(
+  DoRpcOnChannel<std::array<std::uint32_t, 1> >(
       new_channel, AMQP_CHANNEL_OPEN_METHOD, &channel_open, OPEN_OK);
 
-  static const boost::array<boost::uint32_t, 1> CONFIRM_OK = {
-      {AMQP_CONFIRM_SELECT_OK_METHOD}};
+  static const std::array<std::uint32_t, 1> CONFIRM_OK = {
+      AMQP_CONFIRM_SELECT_OK_METHOD};
   amqp_confirm_select_t confirm_select = {};
-  DoRpcOnChannel<boost::array<boost::uint32_t, 1> >(
+  DoRpcOnChannel<std::array<std::uint32_t, 1> >(
       new_channel, AMQP_CONFIRM_SELECT_METHOD, &confirm_select, CONFIRM_OK);
 
   m_channels.at(new_channel) = CS_Open;
@@ -395,18 +393,19 @@ std::vector<amqp_channel_t> Channel::ChannelImpl::GetAllConsumerChannels()
 
 bool Channel::ChannelImpl::CheckForQueuedMessageOnChannel(
     amqp_channel_t channel) const {
-  frame_queue_t::const_iterator it =
-      std::find_if(m_frame_queue.begin(), m_frame_queue.end(),
-                   boost::bind(&Channel::ChannelImpl::is_method_on_channel, _1,
-                               AMQP_BASIC_DELIVER_METHOD, channel));
+  frame_queue_t::const_iterator it = std::find_if(
+      m_frame_queue.begin(), m_frame_queue.end(), [channel](auto &frame) {
+        return ChannelImpl::is_method_on_channel(
+            frame, AMQP_BASIC_DELIVER_METHOD, channel);
+      });
 
   if (it == m_frame_queue.end()) {
     return false;
   }
 
-  it = std::find_if(
-      it + 1, m_frame_queue.end(),
-      boost::bind(&Channel::ChannelImpl::is_on_channel, _1, channel));
+  it = std::find_if(it + 1, m_frame_queue.end(), [channel](auto &frame) {
+    return Channel::ChannelImpl::is_on_channel(frame, channel);
+  });
 
   if (it == m_frame_queue.end()) {
     return false;
@@ -419,9 +418,9 @@ bool Channel::ChannelImpl::CheckForQueuedMessageOnChannel(
   uint64_t body_received = 0;
 
   while (body_received < body_length) {
-    it = std::find_if(
-        it + 1, m_frame_queue.end(),
-        boost::bind(&Channel::ChannelImpl::is_on_channel, _1, channel));
+    it = std::find_if(it + 1, m_frame_queue.end(), [channel](auto &frame) {
+      return Channel::ChannelImpl::is_on_channel(frame, channel);
+    });
 
     if (it == m_frame_queue.end()) {
       return false;
@@ -439,7 +438,7 @@ void Channel::ChannelImpl::AddToFrameQueue(const amqp_frame_t &frame) {
   m_frame_queue.push_back(frame);
 
   if (CheckForQueuedMessageOnChannel(frame.channel)) {
-    boost::array<amqp_channel_t, 1> channel = {{frame.channel}};
+    std::array<amqp_channel_t, 1> channel = {frame.channel};
     Envelope::ptr_t envelope;
     if (!ConsumeMessageOnChannelInner(channel, envelope, -1)) {
       throw std::logic_error(
@@ -451,25 +450,24 @@ void Channel::ChannelImpl::AddToFrameQueue(const amqp_frame_t &frame) {
 }
 
 bool Channel::ChannelImpl::GetNextFrameFromBroker(
-    amqp_frame_t &frame, boost::chrono::microseconds timeout) {
+    amqp_frame_t &frame, std::chrono::microseconds timeout) {
   struct timeval *tvp = NULL;
   struct timeval tv_timeout;
   memset(&tv_timeout, 0, sizeof(tv_timeout));
 
-  if (timeout != boost::chrono::microseconds::max()) {
-    // boost::chrono::seconds.count() returns boost::int_atleast64_t,
+  if (timeout != std::chrono::microseconds::max()) {
+    // std::chrono::seconds.count() returns std::int_atleast64_t,
     // long can be 32 or 64 bit depending on the platform/arch
     // unless the timeout is something absurd cast to long will be ok, but
     // lets guard against the case where someone does something silly
-    assert(
-        boost::chrono::duration_cast<boost::chrono::seconds>(timeout).count() <
-        static_cast<boost::chrono::seconds::rep>(
-            std::numeric_limits<long>::max()));
+    assert(std::chrono::duration_cast<std::chrono::seconds>(timeout).count() <
+           static_cast<std::chrono::seconds::rep>(
+               std::numeric_limits<long>::max()));
 
     tv_timeout.tv_sec = static_cast<long>(
-        boost::chrono::duration_cast<boost::chrono::seconds>(timeout).count());
+        std::chrono::duration_cast<std::chrono::seconds>(timeout).count());
     tv_timeout.tv_usec = static_cast<long>(
-        (timeout - boost::chrono::seconds(tv_timeout.tv_sec)).count());
+        (timeout - std::chrono::seconds(tv_timeout.tv_sec)).count());
 
     tvp = &tv_timeout;
   }
@@ -485,10 +483,11 @@ bool Channel::ChannelImpl::GetNextFrameFromBroker(
 
 bool Channel::ChannelImpl::GetNextFrameOnChannel(
     amqp_channel_t channel, amqp_frame_t &frame,
-    boost::chrono::microseconds timeout) {
+    std::chrono::microseconds timeout) {
   frame_queue_t::iterator it = std::find_if(
-      m_frame_queue.begin(), m_frame_queue.end(),
-      boost::bind(&Channel::ChannelImpl::is_on_channel, _1, channel));
+      m_frame_queue.begin(), m_frame_queue.end(), [channel](auto &frame) {
+        return ChannelImpl::is_on_channel(frame, channel);
+      });
 
   if (m_frame_queue.end() != it) {
     frame = *it;
@@ -503,16 +502,17 @@ bool Channel::ChannelImpl::GetNextFrameOnChannel(
     return true;
   }
 
-  boost::array<amqp_channel_t, 1> channels = {{channel}};
+  std::array<amqp_channel_t, 1> channels = {channel};
   return GetNextFrameFromBrokerOnChannel(channels, frame, timeout);
 }
 
 void Channel::ChannelImpl::MaybeReleaseBuffersOnChannel(
     amqp_channel_t channel) {
   if (m_frame_queue.end() ==
-      std::find_if(
-          m_frame_queue.begin(), m_frame_queue.end(),
-          boost::bind(&Channel::ChannelImpl::is_on_channel, _1, channel))) {
+      std::find_if(m_frame_queue.begin(), m_frame_queue.end(),
+                   [channel](auto &frame) {
+                     return Channel::ChannelImpl::is_on_channel(frame, channel);
+                   })) {
     amqp_maybe_release_buffers_on_channel(m_connection, channel);
   }
 }
@@ -532,9 +532,24 @@ bool bytesEqual(amqp_bytes_t r, amqp_bytes_t l) {
   }
   return false;
 }
+
+std::vector<std::string> splitVersion(const std::string &version) {
+  static char delim = '.';
+  std::vector<std::string> out;
+  std::size_t prev = 0;
+  std::size_t cur = version.find(delim);
+  while (cur != std::string::npos) {
+    out.push_back(version.substr(prev, cur - prev));
+    prev = cur + 1;
+    cur = version.find(delim, prev);
+  }
+  out.push_back(version.substr(prev, cur - prev));
+  return out;
+}
+
 }  // namespace
 
-boost::uint32_t Channel::ChannelImpl::ComputeBrokerVersion(
+std::uint32_t Channel::ChannelImpl::ComputeBrokerVersion(
     amqp_connection_state_t state) {
   const amqp_table_t *properties = amqp_get_server_properties(state);
   const amqp_bytes_t version = amqp_cstring_bytes("version");
@@ -553,17 +568,13 @@ boost::uint32_t Channel::ChannelImpl::ComputeBrokerVersion(
   std::string version_string(
       static_cast<char *>(version_entry->value.value.bytes.bytes),
       version_entry->value.value.bytes.len);
-  std::vector<std::string> version_components;
-  boost::split(version_components, version_string, boost::is_any_of("."));
+  std::vector<std::string> version_components = splitVersion(version_string);
   if (version_components.size() != 3) {
     return 0;
   }
-  boost::uint32_t version_major =
-      boost::lexical_cast<boost::uint32_t>(version_components[0]);
-  boost::uint32_t version_minor =
-      boost::lexical_cast<boost::uint32_t>(version_components[1]);
-  boost::uint32_t version_patch =
-      boost::lexical_cast<boost::uint32_t>(version_components[2]);
+  std::uint32_t version_major = std::stoul(version_components[0]);
+  std::uint32_t version_minor = std::stoul(version_components[1]);
+  std::uint32_t version_patch = std::stoul(version_components[2]);
   return (version_major & 0xFF) << 16 | (version_minor & 0xFF) << 8 |
          (version_patch & 0xFF);
 }
