@@ -28,11 +28,8 @@
  * ***** END LICENSE BLOCK *****
  */
 
-#include <boost/cstdint.hpp>
-#include <boost/make_shared.hpp>
-#include <boost/scoped_ptr.hpp>
-#include <boost/shared_ptr.hpp>
-#include <boost/utility.hpp>
+#include <cstdint>
+#include <memory>
 #include <string>
 
 #include "SimpleAmqpClient/Table.h"
@@ -51,10 +48,10 @@ namespace AmqpClient {
 /**
  * An AMQP BasicMessage
  */
-class SIMPLEAMQPCLIENT_EXPORT BasicMessage : boost::noncopyable {
+class SIMPLEAMQPCLIENT_EXPORT BasicMessage {
  public:
-  /// A shared pointer to BasicMessage
-  typedef boost::shared_ptr<BasicMessage> ptr_t;
+  /// A std::shared_ptr to BasicMessage
+  typedef std::shared_ptr<BasicMessage> ptr_t;
 
   /// With durable queues, messages can be requested to persist or not
   enum delivery_mode_t {
@@ -66,7 +63,7 @@ class SIMPLEAMQPCLIENT_EXPORT BasicMessage : boost::noncopyable {
   /**
    * Create a new empty BasicMessage object
    */
-  static ptr_t Create() { return boost::make_shared<BasicMessage>(); }
+  static ptr_t Create() { return std::make_shared<BasicMessage>(); }
 
   /**
    * Create a new BasicMessage object with given body
@@ -75,7 +72,7 @@ class SIMPLEAMQPCLIENT_EXPORT BasicMessage : boost::noncopyable {
    * @returns a new BasicMessage object
    */
   static ptr_t Create(const std::string& body) {
-    return boost::make_shared<BasicMessage>(body);
+    return std::make_shared<BasicMessage>(body);
   }
 
   /// Construct empty BasicMessage
@@ -84,6 +81,10 @@ class SIMPLEAMQPCLIENT_EXPORT BasicMessage : boost::noncopyable {
   BasicMessage(const std::string& body);
 
  public:
+  // Non-copyable
+  BasicMessage(const BasicMessage &) = delete;
+  BasicMessage &operator=(const BasicMessage &) = delete;
+
   /**
    * Destructor
    */
@@ -154,11 +155,11 @@ class SIMPLEAMQPCLIENT_EXPORT BasicMessage : boost::noncopyable {
   /**
    * Gets the priority property
    */
-  boost::uint8_t Priority() const;
+  std::uint8_t Priority() const;
   /**
    * Sets the priority property
    */
-  void Priority(boost::uint8_t priority);
+  void Priority(std::uint8_t priority);
   /**
    * Determines whether the priority property is set
    */
@@ -239,11 +240,11 @@ class SIMPLEAMQPCLIENT_EXPORT BasicMessage : boost::noncopyable {
   /**
    * Gets the timestamp property
    */
-  boost::uint64_t Timestamp() const;
+  std::uint64_t Timestamp() const;
   /**
    * Sets the timestamp property
    */
-  void Timestamp(boost::uint64_t timestamp);
+  void Timestamp(std::uint64_t timestamp);
   /**
    * Determines whether the timestamp property is set
    */
@@ -342,7 +343,7 @@ class SIMPLEAMQPCLIENT_EXPORT BasicMessage : boost::noncopyable {
  protected:
   struct Impl;
   /// PIMPL idiom
-  boost::scoped_ptr<Impl> m_impl;
+  std::unique_ptr<Impl> m_impl;
 };
 
 }  // namespace AmqpClient
