@@ -30,11 +30,11 @@
 
 #include <amqp.h>
 
-#include <boost/cstdint.hpp>
-#include <boost/shared_ptr.hpp>
-#include <boost/variant/variant.hpp>
 #include <ctime>
+#include <cstdint>
+#include <memory>
 #include <string>
+#include <variant>
 #include <vector>
 
 #include "SimpleAmqpClient/Table.h"
@@ -42,7 +42,7 @@
 namespace AmqpClient {
 namespace Detail {
 
-typedef boost::shared_ptr<amqp_pool_t> amqp_pool_ptr_t;
+typedef std::shared_ptr<amqp_pool_t> amqp_pool_ptr_t;
 
 struct void_t {};
 
@@ -50,10 +50,9 @@ inline bool operator==(const void_t &, const void_t &) { return true; }
 
 typedef std::vector<TableValue> array_t;
 
-typedef boost::variant<void_t, bool, boost::int8_t, boost::int16_t,
-                       boost::int32_t, boost::int64_t, float, double,
-                       std::string, array_t, Table, boost::uint8_t,
-                       boost::uint16_t, boost::uint32_t, boost::uint64_t>
+typedef std::variant<void_t, bool, std::int8_t, std::int16_t, std::int32_t,
+                     std::int64_t, float, double, std::string, array_t, Table,
+                     std::uint8_t, std::uint16_t, std::uint32_t, std::uint64_t>
     value_t;
 
 class TableValueImpl {
@@ -81,22 +80,21 @@ class TableValueImpl {
                                       amqp_pool_t &pool);
 
  public:
-  class generate_field_value
-      : public boost::static_visitor<amqp_field_value_t> {
+  class generate_field_value {
    public:
     explicit generate_field_value(amqp_pool_t &p) : pool(p) {}
     virtual ~generate_field_value() {}
 
     amqp_field_value_t operator()(const void_t) const;
     amqp_field_value_t operator()(const bool value) const;
-    amqp_field_value_t operator()(const boost::uint8_t value) const;
-    amqp_field_value_t operator()(const boost::int8_t value) const;
-    amqp_field_value_t operator()(const boost::uint16_t value) const;
-    amqp_field_value_t operator()(const boost::int16_t value) const;
-    amqp_field_value_t operator()(const boost::uint32_t value) const;
-    amqp_field_value_t operator()(const boost::int32_t value) const;
-    amqp_field_value_t operator()(const boost::uint64_t value) const;
-    amqp_field_value_t operator()(const boost::int64_t value) const;
+    amqp_field_value_t operator()(const std::uint8_t value) const;
+    amqp_field_value_t operator()(const std::int8_t value) const;
+    amqp_field_value_t operator()(const std::uint16_t value) const;
+    amqp_field_value_t operator()(const std::int16_t value) const;
+    amqp_field_value_t operator()(const std::uint32_t value) const;
+    amqp_field_value_t operator()(const std::int32_t value) const;
+    amqp_field_value_t operator()(const std::uint64_t value) const;
+    amqp_field_value_t operator()(const std::int64_t value) const;
     amqp_field_value_t operator()(const float value) const;
     amqp_field_value_t operator()(const double value) const;
     amqp_field_value_t operator()(const std::string &value) const;
