@@ -2,9 +2,9 @@
 #define SIMPLEAMQPCLIENT_BYTES_H
 
 #include <amqp.h>
-#include <boost/utility/string_ref.hpp>
 
 #include <string>
+#include <string_view>
 
 namespace AmqpClient {
 
@@ -15,7 +15,7 @@ amqp_bytes_t StringToBytes(const std::string& str) {
   return ret;
 }
 
-amqp_bytes_t StringRefToBytes(boost::string_ref str) {
+amqp_bytes_t StringRefToBytes(std::string_view str) {
   amqp_bytes_t ret;
   ret.bytes = reinterpret_cast<void*>(const_cast<char*>(str.data()));
   ret.len = str.length();
